refactor(bezier): move _extract_scalar_coeffs to _bezier_sign and reuse in _uniform_sign
Move _extract_scalar_coeffs from _mask to _bezier_sign so both
_uniform_sign and the mask operations share the same coefficient
extraction and weight validation logic.
