+++ src/pantr/bezier/_bezier_sign.py
@@ -85,20 +85,62 @@ def _uniform_sign(bezier: Bezier) -> UniformSign:
         )
 
     if bezier.is_rational:
-        numer: npt.NDArray[np.floating[Any]] = bezier.control_points[..., 0].ravel()
-        weights: npt.NDArray[np.floating[Any]] = bezier.control_points[..., -1].ravel()
-        numer = np.ascontiguousarray(numer, dtype=bezier.dtype)
-        weights = np.ascontiguousarray(weights, dtype=bezier.dtype)
-
-        sign_n = _coeff_sign(numer)
-        if sign_n is UniformSign.MIXED:
-            return UniformSign.MIXED
+        weights = bezier.control_points[..., -1].ravel()
         sign_w = _coeff_sign(weights)
         if sign_w is UniformSign.MIXED:
+            return UniformSign.MIXED
+        # _extract_scalar_coeffs won't raise: rank==1 checked, weights same-sign.
+        sign_n = _coeff_sign(_extract_scalar_coeffs(bezier).ravel())
+        if sign_n is UniformSign.MIXED:
             return UniformSign.MIXED
         # sign(n/w) = sign(n) * sign(w)
         return UniformSign(sign_n * sign_w)
 
-    coeffs: npt.NDArray[np.floating[Any]] = bezier.control_points[..., 0].ravel()
-    coeffs = np.ascontiguousarray(coeffs, dtype=bezier.dtype)
-    return _coeff_sign(coeffs)
+    return _coeff_sign(_extract_scalar_coeffs(bezier).ravel())
+
+
+def _extract_scalar_coeffs(
+    bezier: Bezier,
+) -> npt.NDArray[np.float64]:
+    """Extract scalar Bernstein coefficients from a Bezier as a contiguous N-D float64 array.
+
+    For non-rational Beziers, returns the scalar component ``control_points[..., 0]``.
+
+    For rational Beziers with all weights sharing the same strict sign (all
+    positive or all negative), returns the numerator coefficients
+    ``control_points[..., 0]`` (which are ``w_i * c_i`` in homogeneous form).
+    The zeros of the rational function coincide with the zeros of the
+    numerator when all weights have the same sign.
+
+    Args:
+        bezier (~pantr.bezier.Bezier): A scalar Bezier (``rank == 1``).
+            Rational Beziers are accepted when all weights share the same
+            strict sign.
+
+    Returns:
+        npt.NDArray[np.float64]: Contiguous coefficient array of shape
+        ``(p0+1, p1+1, ...)``.
+
+    Raises:
+        TypeError: If ``bezier`` is rational and the weights do not all share
+            the same strict sign (i.e., some are positive and some negative,
+            or any are zero).
+        ValueError: If ``bezier`` is not scalar (``rank != 1``).
+    """
+    if bezier.rank != 1:
+        raise ValueError(
+            f"_extract_scalar_coeffs requires a scalar Bézier (rank == 1), got rank {bezier.rank}."
+        )
+
+    if bezier.is_rational:
+        weights = bezier.control_points[..., -1]
+        if _coeff_sign(weights.ravel()) is UniformSign.MIXED:
+            raise TypeError(
+                "Rational Bézier operations require all weights to share "
+                "the same strict sign (all positive or all negative)."
+            )
+
+    coeffs: npt.NDArray[np.float64] = np.ascontiguousarray(
+        bezier.control_points[..., 0], dtype=np.float64
+    )
+    return coeffs
